refactor: main함수 내부 UI 구성 요소를 render 함수로 분리
